Make the op execs return the offset of the next pc
Instead of returning the length of the arguments, make the op execs
functions return the offset of the next pc. In most of the case this is
the length of the args plus one (the length of the whole op), except
with zjmp where it can be the jump address.